requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-Login==0.6.3
werkzeug==3.0.1
pytz==2024.2
zxcvbn==4.4.28
Flask-Limiter==3.5.0
redis==5.0.1
Pillow==10.2.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
email-validator==2.1.0.post1
python-dateutil==2.8.2
Flask-Mail==0.9.1